Implement bot core and slash commands handlers.
With this change, we can now begin implementing Slash Commands, and as
long as they're constructed using either the Component annotation or a
Bean, the bot will autowire them and sync them if they aren't in the
global list.

---

Implemented a core package, which provides an EmbedProvider to ensure
consitent embed styling, and a CommandManager, which delegates execution to
command instances. The default embed provider can be easily overriden by
just creating another Bean.

The construction of JDA is handled in JDAConfig under `fvbot.config`.
The implementation was copied from LocketUtils, since that seems to work
rather nicely.

I also implemented two event listeners. One of these - the
CommandSyncHandler - is responsible for comparing local commands against
globally synced commands, and if there's a difference, it does a global
sync. The other listener - SlashCommandInteractionListener - as it's
name suggests, listens for SlashCommand calls and delegates them via the
CommandManager service.